requirements.txt
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy>=1.24.0
torch>=2.0.0 --extra-index-url https://download.pytorch.org/whl/cpu
